:sparkles: feat: Construindo hook e service para abrir votação
